Update TransactionSEnder to typescript
