+++ federator/integrationTest/transfer.js
@@ -5,7 +5,7 @@ const config = require('../config/config.js');
 const abiBridgeV3 = require('../../bridge/abi/Bridge.json');
 const erc20TokenAbi = require('../../bridge/abi/IERC20.json');
 //utils
-const TransactionSender = require('../src/lib/TransactionSender.js');
+const TransactionSender = require('../src/lib/TransactionSender');
 const utils = require('../src/lib/utils.js');
 
 const logger = log4js.getLogger('test');
